fix: navigation to tutorial page done through quick tutorial button
