+++ app/src/main/java/be/scri/ui/screens/about/AboutScreen.kt
@@ -8,26 +8,13 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import be.scri.R
 import be.scri.helpers.AppFlavor
 import be.scri.helpers.FlavorProvider
@@ -37,7 +24,6 @@ import be.scri.ui.common.components.ItemCardContainerWithTitle
 import be.scri.ui.screens.about.AboutUtil.getCommunityList
 import be.scri.ui.screens.about.AboutUtil.getFeedbackAndSupportList
 import be.scri.ui.screens.about.AboutUtil.getLegalListItems
-import be.scri.ui.screens.tutorial.TutorialNavigator
 
 /**
  * The about page of the application with links to the community as well as sub pages for detailed descriptions.
@@ -52,15 +38,6 @@ fun AboutScreen(
     context: Context,
     modifier: Modifier = Modifier,
 ) {
-    var showTutorial by remember { mutableStateOf(false) }
-
-    if (showTutorial) {
-        TutorialNavigator(
-            onTutorialExit = { showTutorial = false },
-        )
-        return
-    }
-
     val isConjugateApp = FlavorProvider.get() == AppFlavor.CONJUGATE
     val scrollState = rememberScrollState()
 
@@ -104,28 +81,6 @@ fun AboutScreen(
                     .verticalScroll(scrollState),
             verticalArrangement = Arrangement.spacedBy(8.dp),
         ) {
-            // Tutorial button
-            Button(
-                onClick = { showTutorial = true },
-                colors =
-                    ButtonDefaults.buttonColors(
-                        containerColor = Color(0xFFF5A623),
-                        contentColor = Color.White,
-                    ),
-                shape = RoundedCornerShape(12.dp),
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 16.dp)
-                        .height(52.dp),
-            ) {
-                Text(
-                    text = "Start full tutorial",
-                    fontSize = 18.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-            }
-
             ItemCardContainerWithTitle(
                 title = stringResource(R.string.i18n_app_about_community_title),
                 cardItemsList = communityList,
